refactor: Kafka 코드 컨벤션 적용
