+++ payment-service/src/main/java/com/palja/payment_service/application/event/dto/request/PaymentCancelFailedEventReq.java
@@ -1,4 +1,4 @@
-package com.palja.payment_service.application.event.dto;
+package com.palja.payment_service.application.event.dto.request;
 
 import lombok.Builder;
 
@@ -7,15 +7,15 @@ import java.time.LocalDateTime;
 import java.util.UUID;
 
 @Builder
-public record PaymentCanceledPayload(
+public record PaymentCancelFailedEventReq(
         UUID paymentId,
         UUID orderId,
         Long userId,
-        BigDecimal canceledAmount,
+        BigDecimal amount,
         String cancelReason,
         String paymentKey,
         String pgResponseCode,
         String pgResponseMessage,
-        LocalDateTime canceledAt
+        LocalDateTime failedAt
 ) {
 }
