+++ payment-service/src/main/java/com/palja/payment_service/application/event/dto/request/PaymentCanceledEventReq.java
@@ -1,4 +1,4 @@
-package com.palja.payment_service.application.event.dto;
+package com.palja.payment_service.application.event.dto.request;
 
 import lombok.Builder;
 
@@ -7,14 +7,15 @@ import java.time.LocalDateTime;
 import java.util.UUID;
 
 @Builder
-public record PaymentApprovedPayload(
+public record PaymentCanceledEventReq(
         UUID paymentId,
         UUID orderId,
         Long userId,
-        BigDecimal paidAmount,
+        BigDecimal canceledAmount,
+        String cancelReason,
         String paymentKey,
         String pgResponseCode,
         String pgResponseMessage,
-        LocalDateTime approvedAt
+        LocalDateTime canceledAt
 ) {
 }
