+++ payment-service/src/main/java/com/palja/payment_service/application/service/PaymentSagaService.java
@@ -1,7 +1,7 @@
 package com.palja.payment_service.application.service;
 
-import com.palja.payment_service.infrastructure.saga.dto.request.PaymentCancelEventReq;
-import com.palja.payment_service.infrastructure.saga.dto.request.PaymentCreateEventReq;
+import com.palja.payment_service.application.event.dto.request.PaymentCancelEventReq;
+import com.palja.payment_service.application.event.dto.request.PaymentCreateEventReq;
 
 import java.util.UUID;
