+++ payment-service/src/main/java/com/palja/payment_service/application/service/impl/PaymentSagaServiceImpl.java
@@ -10,8 +10,8 @@ import com.palja.payment_service.domain.repository.PaymentLogRepository;
 import com.palja.payment_service.domain.repository.PaymentRepository;
 import com.palja.payment_service.domain.vo.PaymentStatus;
 import com.palja.payment_service.exception.PaymentErrorCode;
-import com.palja.payment_service.infrastructure.saga.dto.request.PaymentCancelEventReq;
-import com.palja.payment_service.infrastructure.saga.dto.request.PaymentCreateEventReq;
+import com.palja.payment_service.application.event.dto.request.PaymentCancelEventReq;
+import com.palja.payment_service.application.event.dto.request.PaymentCreateEventReq;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
