+++ payment-service/src/main/java/com/palja/payment_service/infrastructure/config/KafkaProducerConfig.java
@@ -1,5 +1,7 @@
 package com.palja.payment_service.infrastructure.config;
 
+import com.palja.common.interceptor.KafkaProducerInterceptor;
+import io.micrometer.tracing.Tracer;
 import org.apache.kafka.clients.producer.ProducerConfig;
 import org.apache.kafka.common.serialization.StringSerializer;
 import org.springframework.beans.factory.annotation.Value;
@@ -8,6 +10,7 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.kafka.core.DefaultKafkaProducerFactory;
 import org.springframework.kafka.core.KafkaTemplate;
 import org.springframework.kafka.core.ProducerFactory;
+import org.springframework.kafka.support.serializer.JsonSerializer;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -19,25 +22,52 @@ public class KafkaProducerConfig {
     private String bootstrapServers;
 
     @Bean
-    public ProducerFactory<String, String> producerFactory() {
-        Map<String, Object> configs = new HashMap<>();
+    public ProducerFactory<String, String> outboxProducerFactory() {
+        Map<String, Object> config = new HashMap<>();
 
-        configs.put(ProducerConfig.BOOTSTRAP_SERVERS_CONFIG, bootstrapServers);
-        configs.put(ProducerConfig.KEY_SERIALIZER_CLASS_CONFIG, StringSerializer.class);
-        configs.put(ProducerConfig.VALUE_SERIALIZER_CLASS_CONFIG, StringSerializer.class);
+        config.put(ProducerConfig.BOOTSTRAP_SERVERS_CONFIG, bootstrapServers);
+        config.put(ProducerConfig.KEY_SERIALIZER_CLASS_CONFIG, StringSerializer.class);
+        config.put(ProducerConfig.VALUE_SERIALIZER_CLASS_CONFIG, StringSerializer.class);
 
-        configs.put(ProducerConfig.ACKS_CONFIG, "all");
-        configs.put(ProducerConfig.ENABLE_IDEMPOTENCE_CONFIG, true);
-        configs.put(ProducerConfig.RETRIES_CONFIG, Integer.MAX_VALUE);
+        config.put(ProducerConfig.ACKS_CONFIG, "all");
+        config.put(ProducerConfig.ENABLE_IDEMPOTENCE_CONFIG, true);
+        config.put(ProducerConfig.RETRIES_CONFIG, Integer.MAX_VALUE);
+        config.put(ProducerConfig.MAX_IN_FLIGHT_REQUESTS_PER_CONNECTION, 5);
+        config.put(ProducerConfig.COMPRESSION_TYPE_CONFIG, "snappy");
 
-        configs.put(ProducerConfig.MAX_IN_FLIGHT_REQUESTS_PER_CONNECTION, 5);
-        configs.put(ProducerConfig.COMPRESSION_TYPE_CONFIG, "snappy");
+        return new DefaultKafkaProducerFactory<>(config);
+    }
 
-        return new DefaultKafkaProducerFactory<>(configs);
+    @Bean(name = "outboxKafkaTemplate")
+    public KafkaTemplate<String, String> outboxKafkaTemplate() {
+        return new KafkaTemplate<>(outboxProducerFactory());
     }
 
     @Bean
-    public KafkaTemplate<String, String> kafkaTemplate() {
-        return new KafkaTemplate<>(producerFactory());
+    public KafkaProducerInterceptor<Object> kafkaProducerInterceptor(Tracer tracer) {
+        return new KafkaProducerInterceptor<>(tracer);
+    }
+
+    @Bean
+    public ProducerFactory<String, Object> sagaProducerFactory() {
+        Map<String, Object> config = new HashMap<>();
+
+        config.put(ProducerConfig.BOOTSTRAP_SERVERS_CONFIG, bootstrapServers);
+        config.put(ProducerConfig.KEY_SERIALIZER_CLASS_CONFIG, StringSerializer.class);
+        config.put(ProducerConfig.VALUE_SERIALIZER_CLASS_CONFIG, JsonSerializer.class);
+
+        config.put(ProducerConfig.ACKS_CONFIG, "all");
+        config.put(ProducerConfig.ENABLE_IDEMPOTENCE_CONFIG, true);
+
+        return new DefaultKafkaProducerFactory<>(config);
+    }
+
+    @Bean(name = "sagaKafkaTemplate")
+    public KafkaTemplate<String, Object> sagaKafkaTemplate(
+            KafkaProducerInterceptor<Object> interceptor
+    ) {
+        KafkaTemplate<String, Object> template = new KafkaTemplate<>(sagaProducerFactory());
+        template.setProducerInterceptor(interceptor);
+        return template;
     }
 }
