+++ payment-service/src/main/java/com/palja/payment_service/infrastructure/external/kafka/producer/PaymentSagaReplyProducer.java
@@ -1,7 +1,7 @@
-package com.palja.payment_service.infrastructure.saga.producer;
+package com.palja.payment_service.infrastructure.external.kafka.producer;
 
-import com.palja.payment_service.infrastructure.saga.dto.response.PaymentCreateEventRes;
-import com.palja.payment_service.infrastructure.saga.topics.OrderSagaKafkaTopics;
+import com.palja.payment_service.application.event.dto.response.PaymentCreateEventRes;
+import com.palja.payment_service.infrastructure.external.kafka.KafkaTopics;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Qualifier;
@@ -17,13 +17,13 @@ public class PaymentSagaReplyProducer {
     private final KafkaTemplate<String, Object> kafkaTemplate;
 
     public void publishCreateSuccess(PaymentCreateEventRes res) {
-        sendSync(OrderSagaKafkaTopics.PAYMENT_CREATE_SUCCESS, res.getSagaId().toString(), res);
+        sendSync(KafkaTopics.PAYMENT_CREATE_SUCCESS, res.getSagaId().toString(), res);
         log.info("saga 결제 응답 생성 성공: sagaId={}, orderId={}, paymentId={}",
                 res.getSagaId(), res.getOrderId(), res.getPaymentId());
     }
 
     public void publishCreateFailure(PaymentCreateEventRes res) {
-        sendSync(OrderSagaKafkaTopics.PAYMENT_CREATE_FAILURE, res.getSagaId().toString(), res);
+        sendSync(KafkaTopics.PAYMENT_CREATE_FAILURE, res.getSagaId().toString(), res);
         log.info("saga 결제 응답 생성 실패: sagaId={}, orderId={}",
                 res.getSagaId(), res.getOrderId());
     }
