--- payment-service/src/main/java/com/palja/payment_service/infrastructure/saga/config/SagaKafkaProducerConfig.java
@@ -1,48 +1,0 @@
-package com.palja.payment_service.infrastructure.saga.config;
-
-import com.palja.common.interceptor.KafkaProducerInterceptor;
-import io.micrometer.tracing.Tracer;
-import org.apache.kafka.clients.producer.ProducerConfig;
-import org.apache.kafka.common.serialization.StringSerializer;
-import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.kafka.core.DefaultKafkaProducerFactory;
-import org.springframework.kafka.core.KafkaTemplate;
-import org.springframework.kafka.core.ProducerFactory;
-import org.springframework.kafka.support.serializer.JsonSerializer;
-
-import java.util.HashMap;
-import java.util.Map;
-
-@Configuration
-public class SagaKafkaProducerConfig {
-
-    @Value("${spring.kafka.bootstrap-servers}")
-    private String bootstrapServers;
-
-    @Bean
-    public KafkaProducerInterceptor<Object> sagaKafkaProducerInterceptor(Tracer tracer) {
-        return new KafkaProducerInterceptor<>(tracer);
-    }
-
-    @Bean
-    public ProducerFactory<String, Object> sagaProducerFactory() {
-        Map<String, Object> config = new HashMap<>();
-        config.put(ProducerConfig.BOOTSTRAP_SERVERS_CONFIG, bootstrapServers);
-        config.put(ProducerConfig.KEY_SERIALIZER_CLASS_CONFIG, StringSerializer.class);
-        config.put(ProducerConfig.VALUE_SERIALIZER_CLASS_CONFIG, JsonSerializer.class);
-        config.put(ProducerConfig.ENABLE_IDEMPOTENCE_CONFIG, true);
-        return new DefaultKafkaProducerFactory<>(config);
-    }
-
-    @Bean(name = "sagaKafkaTemplate")
-    public KafkaTemplate<String, Object> sagaKafkaTemplate(
-            @Qualifier("sagaKafkaProducerInterceptor") KafkaProducerInterceptor<Object> interceptor
-    ) {
-        KafkaTemplate<String, Object> template = new KafkaTemplate<>(sagaProducerFactory());
-        template.setProducerInterceptor(interceptor);
-        return template;
-    }
-}
